+++ homeassistant/components/blue_current/strings.json
