+++ homeassistant/components/blue_current/strings.sensor.json
@@ -1,13 +1,13 @@
 {
   "state": {
-    "bluecurrent__activity": {
+    "blue_current__activity": {
       "available": "Available",
       "charging": "Charging",
       "unavailable": "Unavailable",
       "error": "Error",
       "offline": "Offline"
     },
-    "bluecurrent__vehicle_status": {
+    "blue_current__vehicle_status": {
       "standby": "Standby",
       "vehicle_detected": "Detected",
       "ready": "Ready",
