+++ homeassistant/components/blue_current/translations/en.json
