+++ mypy.ini
@@ -570,7 +570,7 @@ disallow_untyped_defs = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.bluecurrent.*]
+[mypy-homeassistant.components.blue_current.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_subclassing_any = true
