+++ requirements_all.txt
@@ -552,7 +552,7 @@ blinkpy==0.22.4
 # homeassistant.components.bitcoin
 blockchain==1.4.4
 
-# homeassistant.components.bluecurrent
+# homeassistant.components.blue_current
 bluecurrent-api==1.0.1
 
 # homeassistant.components.bluemaestro
