+++ requirements_test_all.txt
@@ -468,7 +468,7 @@ blebox-uniapi==2.2.0
 # homeassistant.components.blink
 blinkpy==0.22.4
 
-# homeassistant.components.bluecurrent
+# homeassistant.components.blue_current
 bluecurrent-api==1.0.1
 
 # homeassistant.components.bluemaestro
